requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
numpy==2.2.4
pillow==11.1.0
pydantic==2.10.6
pandas==2.2.3
matplotlib==3.10.1
scikit-learn==1.6.1  # For KNN
pydantic-settings==2.8.1
python-multipart==0.0.20